+++ projects/net.splitcells.dem.api/src/main/java/net/splitcells/dem/data/set/AppendableSet.java
@@ -18,17 +18,22 @@ package net.splitcells.dem.data.set;
 import net.splitcells.dem.lang.annotations.JavaLegacyBody;
 import net.splitcells.dem.lang.annotations.ReturnsThis;
 
-public interface SetWA<T> {
+/**
+ * This is a set interface that contains only methods to add values.
+ *
+ * @param <T> The type of values contained in this set.
+ */
+public interface AppendableSet<T> {
 
 	/**
 	 * TODO Create method which adds one if not present and otherwise throws exception.
 	 */
 	@ReturnsThis
-	<R extends SetWA<T>> R add(T value);
+	<R extends AppendableSet<T>> R add(T value);
 
 	@SuppressWarnings("unchecked")
 	@ReturnsThis
-	default <R extends SetWA<T>> R addAll(T... values) {
+	default <R extends AppendableSet<T>> R addAll(T... values) {
 		for (int i = 0; i < values.length; ++i) {
 			add(values[i]);
 		}
@@ -38,30 +43,30 @@ public interface SetWA<T> {
 	@JavaLegacyBody
 	@SuppressWarnings("unchecked")
 	@ReturnsThis
-	default <R extends SetWA<T>> R addAll(java.util.Collection<T> values) {
+	default <R extends AppendableSet<T>> R addAll(java.util.Collection<T> values) {
 		values.forEach(value -> add(value));
 		return (R) this;
 	}
 
-	static <B> SetWA<B> extend(final B prefix, final SetWA<B> sender) {
-		return new SetWA<B>() {
+	static <B> AppendableSet<B> extend(final B prefix, final AppendableSet<B> sender) {
+		return new AppendableSet<B>() {
 			@SuppressWarnings("unchecked")
 			@ReturnsThis
-			public <R extends SetWA<B>> R add(B arg) {
+			public <R extends AppendableSet<B>> R add(B arg) {
 				sender.addAll(prefix, arg);
 				return (R) this;
 			}
 		};
 	}
 
-	static <A, B> SetWA<B> extend(final B first_prefix, final B defaultPrefix, final SetWA<B> sender) {
-		return new SetWA<B>() {
+	static <A, B> AppendableSet<B> extend(final B first_prefix, final B defaultPrefix, final AppendableSet<B> sender) {
+		return new AppendableSet<B>() {
 
 			private boolean wasFirstAdded = false;
 
 			@SuppressWarnings("unchecked")
 			@ReturnsThis
-			public <R extends SetWA<B>> R add(B arg) {
+			public <R extends AppendableSet<B>> R add(B arg) {
 				if (wasFirstAdded) {
 					sender.addAll(defaultPrefix, arg);
 				} else {