+++ projects/net.splitcells.dem.api/src/main/java/net/splitcells/dem/resource/communication/Sender.java
@@ -15,12 +15,12 @@
  */
 package net.splitcells.dem.resource.communication;
 
-import net.splitcells.dem.data.set.SetWA;
-import net.splitcells.dem.data.set.list.ListWA;
+import net.splitcells.dem.data.set.AppendableSet;
+import net.splitcells.dem.data.set.list.AppendableList;
 import net.splitcells.dem.environment.resource.Resource;
 import net.splitcells.dem.lang.annotations.JavaLegacyBody;
 
-public interface Sender<T> extends ListWA<T>, Resource {
+public interface Sender<T> extends AppendableList<T>, Resource {
 
     @JavaLegacyBody
     static Sender<String> stringSender(java.io.OutputStream output) {
@@ -38,7 +38,7 @@ public interface Sender<T> extends ListWA<T>, Resource {
             private java.io.PrintWriter printer = new java.io.PrintWriter(output);
 
             @Override
-            public <R extends ListWA<String>> R append(String arg) {
+            public <R extends AppendableList<String>> R append(String arg) {
                 /* The new line ending added by println at the end of the argument is
                  * dependent on the current operation system.
                  * To make the program portable, the line ending is done via a constant.
@@ -54,7 +54,7 @@ public interface Sender<T> extends ListWA<T>, Resource {
             }
 
             @Override
-            public <R extends SetWA<String>> R add(String value) {
+            public <R extends AppendableSet<String>> R add(String value) {
                 throw new UnsupportedOperationException();
             }
 
@@ -73,7 +73,7 @@ public interface Sender<T> extends ListWA<T>, Resource {
             private java.io.PrintWriter printer = new java.io.PrintWriter(output);
 
             @Override
-            public <R extends ListWA<String>> R append(String arg) {
+            public <R extends AppendableList<String>> R append(String arg) {
                 printer.println(arg);
                 printer.flush();
                 return (R) this;
@@ -85,7 +85,7 @@ public interface Sender<T> extends ListWA<T>, Resource {
             }
 
             @Override
-            public <R extends SetWA<String>> R add(String value) {
+            public <R extends AppendableSet<String>> R add(String value) {
                 throw new UnsupportedOperationException();
             }
 
@@ -110,13 +110,13 @@ public interface Sender<T> extends ListWA<T>, Resource {
 
             @SuppressWarnings("unchecked")
             @Override
-            public <R extends ListWA<String>> R append(String value) {
+            public <R extends AppendableList<String>> R append(String value) {
                 sender.append(prefix + value + suffix);
                 return (R) this;
             }
 
             @Override
-            public <R extends SetWA<String>> R add(String value) {
+            public <R extends AppendableSet<String>> R add(String value) {
                 throw new UnsupportedOperationException();
             }
 
@@ -140,7 +140,7 @@ public interface Sender<T> extends ListWA<T>, Resource {
 
             @SuppressWarnings("unchecked")
             @Override
-            public <R extends ListWA<String>> R append(String value) {
+            public <R extends AppendableList<String>> R append(String value) {
                 if (is_first) {
                     sender.append(first_prefix + value + first_suffix);
                 } else {
@@ -150,7 +150,7 @@ public interface Sender<T> extends ListWA<T>, Resource {
             }
 
             @Override
-            public <R extends SetWA<String>> R add(String value) {
+            public <R extends AppendableSet<String>> R add(String value) {
                 throw new UnsupportedOperationException();
             }
         };