+++ projects/net.splitcells.dem.api/src/main/java/net/splitcells/dem/resource/communication/log/Log.java
@@ -15,7 +15,7 @@
  */
 package net.splitcells.dem.resource.communication.log;
 
-import net.splitcells.dem.data.set.list.ListWA;
+import net.splitcells.dem.data.set.list.AppendableList;
 import net.splitcells.dem.lang.annotations.JavaLegacyBody;
 import net.splitcells.dem.lang.annotations.ReturnsThis;
 import net.splitcells.dem.lang.dom.Domable;
@@ -39,7 +39,7 @@ import static net.splitcells.dem.utils.StringUtils.throwableToString;
  * In other words, this may be should be implemented as a general
  * functionality in order to provide one message one line logs.</p>
  */
-public interface Log extends ListWA<LogMessage<Tree>> {
+public interface Log extends AppendableList<LogMessage<Tree>> {
 
     default Log append(String name) {
         return append(logMessage(tree(name), NO_CONTEXT, LogLevel.DEBUG));