+++ projects/net.splitcells.dem.api/src/main/java/net/splitcells/dem/resource/communication/log/LoggerImpl.java
@@ -15,7 +15,7 @@
  */
 package net.splitcells.dem.resource.communication.log;
 
-import net.splitcells.dem.data.set.list.ListWA;
+import net.splitcells.dem.data.set.list.AppendableList;
 import net.splitcells.dem.lang.tree.Tree;
 import net.splitcells.dem.resource.communication.Sender;
 
@@ -41,7 +41,7 @@ public class LoggerImpl implements Logger {
     }
 
     @Override
-    public <R extends ListWA<LogMessage<Tree>>> R append(LogMessage<Tree> arg) {
+    public <R extends AppendableList<LogMessage<Tree>>> R append(LogMessage<Tree> arg) {
         if (messageFilter.test(arg)) {
             print(output, arg.content());
         }